4.13.0-rc3

+++ EMMA_iOS.xcframework/ios-arm64_x86_64-simulator/EMMA_iOS.framework/Headers/EMMA_iOS-Swift.h
@@ -319,8 +319,8 @@ SWIFT_CLASS("_TtC8EMMA_iOS18EMMAAdBallCampaign")
 
 SWIFT_CLASS("_TtC8EMMA_iOS16EMMABannerParams")
 @interface EMMABannerParams : NSObject
-@property (nonatomic) NSInteger offsetTop;
-@property (nonatomic) NSInteger offsetBottom;
+@property (nonatomic) double offsetTop;
+@property (nonatomic) double offsetBottom;
 - (nonnull instancetype)init OBJC_DESIGNATED_INITIALIZER;
 @end
 
@@ -839,8 +839,8 @@ SWIFT_CLASS("_TtC8EMMA_iOS18EMMAAdBallCampaign")
 
 SWIFT_CLASS("_TtC8EMMA_iOS16EMMABannerParams")
 @interface EMMABannerParams : NSObject
-@property (nonatomic) NSInteger offsetTop;
-@property (nonatomic) NSInteger offsetBottom;
+@property (nonatomic) double offsetTop;
+@property (nonatomic) double offsetBottom;
 - (nonnull instancetype)init OBJC_DESIGNATED_INITIALIZER;
 @end
